Wait up to 5 seconds for keytool exit in ExpiringKeystoreHealthCheckTest
* Change ExpiringKeystoreHealthCheckTest so that it waits up to five
  seconds when creating a new keystore
* Log an INFO message if process exits within 5 seconds, otherwise WARN
* Update debug logging statements to clarify what is stdout and stderr
  of the keytool processes
* Remove redundant "throws Exception" from tests

Background for this update:

While investigating the failing tests in #104 I temporarily changed the
test code to copy each keystore for each test from the temp directory to
a permanent location for examination. But immediately after the
waitForExit call in createTemporaryKeystore, the Files#copy call was
failing b/c with a FileNotFoundException! So, I logged the exit code and
found that it was an empty Optional, which is what waitForExit returns
if the timeout occurs before the process exits. I therefore increased
the timeout period to 5 seconds to be safe. It seems to take about a
second for it to actually exit, so the original 500 milliseconds was
definitely not enough time. But the tests still passed, presumably
because the process exited and the keystore was flushed to disk between
the call to waitForExit and then the two reads of the process std IO and
std error before the method returned.