Change MonitoredJobs to convert delays to nanoseconds when scheduling
This avoids problems if callers want to schedule more granular than one
second. For example, with initial delay of 100 milliseconds and interval
delay of 200 milliseconds, converting to seconds results in both values
being zero in the original code. By converting the delays to nanoseconds
instead, we ensure this problem won't happen.

Fixes #173