Change MongoHealthCheck to work with 4.x driver
This is a breaking API change, since the DB class no longer exists
in the Mongo 4.x driver. It was  replaced by MongoDatabase. Thus,
the constructor had to change. The MongoDatabase also does not
contain as much information, e.g. you cannot obtain the host and port
information as you could with DB, which provided a getMongo() method
from which you could obtain a ServerAddress in order to find the host
and port. We decided that information does not add enough value to
require callers to include a host and port, though we could re-evauate
that decision later.

This commit, therefore, changes MongoHealthCheck to work with the
4.x driver. The health check has been using the "dbStats" command which
"returns storage statistics for a given database". The problem with this
is that, from the docs, "The time required to run the command depends on
the total size of the database. Because the command must touch all data
files, the command may take several seconds to run." Health checks
should definitely not run any commands that could cause performance
problems or take seconds to execute.

This commit changes the MongoHealthCheck to use the ping command, which
"is a no-op used to test whether a server is responding to commands",
and it "will return immediately even if the server is write-locked". This
is much better for a health check, and should obviously be much faster and
less resource intensive on Mongo.

Closes #233