Fix flaky test in SystemExecutionerTest (#604)
* Update shouldExitBeforeGivenWaitTime_WhenWaitingThreadInterrupted so
that the didExit is a condition of the await, in addition to requiring
the Futures to be done. This eliminates various concurrency problems
such as race conditions, scheduling timing, etc. I also changed the poll
interval from the default (100 millis) to 25 millis so that the
condition matches more quickly than it otherwise would. Sometimes we
were seeing over 200 millis because it would need to poll twice.
* Added a "fudge factor" when making assertions on elapsed time in
shouldExitBeforeGivenWaitTime_WhenWaitingThreadInterrupted and
shouldExitWithWaitTime, again to reduce the change of timing issues.
* Make several of the tests use RepeatedTest with 3 repetitions. This is
mainly to be more confident we've caught all the concurrency issues
(assuming they always pass from now on).

Fixes #594